update ansible provisioning script

+++ ansible-provision.sh
@@ -1,7 +1,17 @@
 #!/bin/bash
 
+echo "Udating packages"
 apt-get update -y
 
-apt-get install -y ansible
+#dpkg -s ansible &>/dev/null || {
+#  echo "ansible is not installed - installling now..."
+#  apt-get install -y ansible
+#}
 
+if dpkg -s ansible > /dev/null 2>&1; then
+  echo "ansible already installed, bypassing..."
+else
+  echo "ansible not installed, installing now..."
+  apt-get install -y ansible
+fi
